Code cleanup - step 2
Split a huge sources.py file into smaller ones
